Implemented the base snake game

+++ NiosDisplay_App/types.c
@@ -4,6 +4,6 @@ bool point_t_equals(const point_t rhs, const point_t lhs) {
     return (rhs.x == lhs.x) && (rhs.y == lhs.y);
 }
 
-bool is_between(u8 min, u8 max, u8 val) {
+bool is_between(uint16_t min, uint16_t max, uint16_t val) {
     return ((val < max) && (val > min)) || val == min;
 }
